requirements.txt
kodi.tools==1.1.0
Kodistubs==21.0.0
pytest==6.2.5
script.module.akl==1.2.0
requests==2.22.0
flake8==4.0.1
